Refractored the kilkari report services as per the ETL. Updated models and corresponding Dao/Service Impls.

+++ NMSReportingSuite/src/main/java/com/beehyv/nmsreporting/dao/SubcenterDao.java
@@ -11,8 +11,7 @@ public interface SubcenterDao {
 
     public Subcenter findBySubcenterId(Integer SubcenterId);
 
-    Subcenter findByLocationId(Long stateId);
-
+    Subcenter findByLocationId(Integer stateId);
 
     public List<Subcenter> findByName(String SubcenterName);
