+++ NMSReportingSuite/src/main/java/com/beehyv/nmsreporting/entity/KilkariMessageListenershipReportDto.java
@@ -9,16 +9,16 @@ package com.beehyv.nmsreporting.entity;
 public class KilkariMessageListenershipReportDto {
 
     private int id;
-    private Integer locationId;
+    private Long locationId;
     private String locationType;
     private String locationName;
-    private Integer totalBeneficiariesCalled;
-    private Integer beneficiariesAnsweredAtleastOnce;
-    private Integer beneficiariesAnsweredMoreThan75;
-    private Integer beneficiariesAnswered50To75;
-    private Integer beneficiariesAnswered25To50;
-    private Integer beneficiariesAnswered1To25;
-    private Integer beneficiariesAnsweredNoCalls;
+    private Long totalBeneficiariesCalled;
+    private Long beneficiariesAnsweredAtleastOnce;
+    private Long beneficiariesAnsweredMoreThan75;
+    private Long beneficiariesAnswered50To75;
+    private Long beneficiariesAnswered25To50;
+    private Long beneficiariesAnswered1To25;
+    private Long beneficiariesAnsweredNoCalls;
 
     public int getId() {
         return id;
@@ -28,11 +28,11 @@ public class KilkariMessageListenershipReportDto {
         this.id = id;
     }
 
-    public Integer getLocationId() {
+    public Long getLocationId() {
         return locationId;
     }
 
-    public void setLocationId(Integer locationId) {
+    public void setLocationId(Long locationId) {
         this.locationId = locationId;
     }
 
@@ -52,59 +52,59 @@ public class KilkariMessageListenershipReportDto {
         this.locationName = locationName;
     }
 
-    public Integer getTotalBeneficiariesCalled() {
+    public Long getTotalBeneficiariesCalled() {
         return totalBeneficiariesCalled;
     }
 
-    public void setTotalBeneficiariesCalled(Integer totalBeneficiariesCalled) {
+    public void setTotalBeneficiariesCalled(Long totalBeneficiariesCalled) {
         this.totalBeneficiariesCalled = totalBeneficiariesCalled;
     }
 
-    public Integer getBeneficiariesAnsweredAtleastOnce() {
+    public Long getBeneficiariesAnsweredAtleastOnce() {
         return beneficiariesAnsweredAtleastOnce;
     }
 
-    public void setBeneficiariesAnsweredAtleastOnce(Integer beneficiariesAnsweredAtleastOnce) {
+    public void setBeneficiariesAnsweredAtleastOnce(Long beneficiariesAnsweredAtleastOnce) {
         this.beneficiariesAnsweredAtleastOnce = beneficiariesAnsweredAtleastOnce;
     }
 
-    public Integer getBeneficiariesAnsweredMoreThan75() {
+    public Long getBeneficiariesAnsweredMoreThan75() {
         return beneficiariesAnsweredMoreThan75;
     }
 
-    public void setBeneficiariesAnsweredMoreThan75(Integer beneficiariesAnsweredMoreThan75) {
+    public void setBeneficiariesAnsweredMoreThan75(Long beneficiariesAnsweredMoreThan75) {
         this.beneficiariesAnsweredMoreThan75 = beneficiariesAnsweredMoreThan75;
     }
 
-    public Integer getBeneficiariesAnswered50To75() {
+    public Long getBeneficiariesAnswered50To75() {
         return beneficiariesAnswered50To75;
     }
 
-    public void setBeneficiariesAnswered50To75(Integer beneficiariesAnswered50To75) {
+    public void setBeneficiariesAnswered50To75(Long beneficiariesAnswered50To75) {
         this.beneficiariesAnswered50To75 = beneficiariesAnswered50To75;
     }
 
-    public Integer getBeneficiariesAnswered25To50() {
+    public Long getBeneficiariesAnswered25To50() {
         return beneficiariesAnswered25To50;
     }
 
-    public void setBeneficiariesAnswered25To50(Integer beneficiariesAnswered25To50) {
+    public void setBeneficiariesAnswered25To50(Long beneficiariesAnswered25To50) {
         this.beneficiariesAnswered25To50 = beneficiariesAnswered25To50;
     }
 
-    public Integer getBeneficiariesAnswered1To25() {
+    public Long getBeneficiariesAnswered1To25() {
         return beneficiariesAnswered1To25;
     }
 
-    public void setBeneficiariesAnswered1To25(Integer beneficiariesAnswered1To25) {
+    public void setBeneficiariesAnswered1To25(Long beneficiariesAnswered1To25) {
         this.beneficiariesAnswered1To25 = beneficiariesAnswered1To25;
     }
 
-    public Integer getBeneficiariesAnsweredNoCalls() {
+    public Long getBeneficiariesAnsweredNoCalls() {
         return beneficiariesAnsweredNoCalls;
     }
 
-    public void setBeneficiariesAnsweredNoCalls(Integer beneficiariesAnsweredNoCalls) {
+    public void setBeneficiariesAnsweredNoCalls(Long beneficiariesAnsweredNoCalls) {
         this.beneficiariesAnsweredNoCalls = beneficiariesAnsweredNoCalls;
     }
 }
